requirements.txt
Flask>=3.0
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
scipy>=1.11
